Try to add DMAP support to XMHF
* test

* test

* test

* test

* Add IOMMU code

* Fix ACPI issue

* Moving common dmap-* files out

* test

* test

* test

* Clean the x86 code

* Fix most of the comments

* Fixed most issues

* Remove eflags and esp from struct regs

* Fix CR0

* Remove optimization, TrustVisor can work

* Also remove ESP and optimization in x86

* Write VCPU_reg_get() and VCPU_reg_set()

* Disable DMAP in configure

* Remove things cleanly

* PAGE_MASK_1GB -> PAGE_MASK_1G

+++ xmhf/src/xmhf-core/xmhf-runtime/xmhf-baseplatform/arch/x86_64/bplt-x86_64-acpi.c
@@ -73,7 +73,7 @@ static u32 _acpi_computetablechecksum(uintptr_t spaddr, uintptr_t size){
 //get the physical address of the root system description pointer (rsdp)
 //return 0 in case of error (ACPI RSDP not found) else the absolute physical
 //memory address of the RSDP
-u32 xmhf_baseplatform_arch_x86_64_acpi_getRSDP(ACPI_RSDP *rsdp){
+uintptr_t xmhf_baseplatform_arch_x86_64_acpi_getRSDP(ACPI_RSDP *rsdp){
   u16 ebdaseg;
   uintptr_t ebdaphys;
   uintptr_t i;
@@ -98,7 +98,7 @@ u32 xmhf_baseplatform_arch_x86_64_acpi_getRSDP(ACPI_RSDP *rsdp){
 
 	//found RSDP?  
   if(found)
-    return (u32)(ebdaphys+i);
+    return (ebdaphys+i);
   
   //nope, search within BIOS areas 0xE0000 to 0xFFFFF
   for(i=0xE0000; i < (0xFFFFF-8); i+=16){
